requirements.txt
Django==4.2.8
djangorestframework>=3.14
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.2
Pillow>=9.0.0
django-cors-headers
